+++ example/android/app/src/main/java/com/example/clevertap_plugin_example/MyApplication.java
@@ -4,10 +4,8 @@ import com.clevertap.android.pushtemplates.PushTemplateNotificationHandler;
 import com.clevertap.android.pushtemplates.TemplateRenderer;
 import com.clevertap.android.sdk.ActivityLifecycleCallback;
 import com.clevertap.android.sdk.CleverTapAPI;
-import com.clevertap.android.sdk.interfaces.NotificationHandler;
 import com.clevertap.clevertap_plugin.CleverTapApplication;
-
-import io.flutter.app.FlutterApplication;
+import com.clevertap.clevertap_plugin.ClevertapCustomTemplates;
 
 public class MyApplication extends CleverTapApplication {
 
@@ -16,6 +14,7 @@ public class MyApplication extends CleverTapApplication {
         CleverTapAPI.setDebugLevel(3);
         TemplateRenderer.setDebugLevel(3);
         CleverTapAPI.setNotificationHandler(new PushTemplateNotificationHandler());
+        ClevertapCustomTemplates.registerCustomTemplates(this, "templates.json");
         ActivityLifecycleCallback.register(this);
         super.onCreate();
     }