+++ lib/src/typedefs.dart
@@ -21,10 +21,20 @@ typedef void CleverTapPushClickedPayloadReceivedHandler(
     Map<String, dynamic> map);
 typedef void CleverTapPushPermissionResponseReceivedHandler(bool accepted);
 typedef void CleverTapOnVariablesChangedHandler(Map<String, dynamic> variables);
+typedef void CleverTapOnOneTimeVariablesChangedHandler(Map<String, dynamic> variables);
 typedef void CleverTapOnValueChangedHandler(Map<String, dynamic> variable);
+typedef void CleverTapOnVariablesChangedAndNoDownloadsPendingHandler(Map<String, dynamic> variables);
+typedef void CleverTapOnceVariablesChangedAndNoDownloadsPendingHandler(Map<String, dynamic> variables);
+typedef void CleverTapOnFileValueChangedHandler(Map<String, dynamic> variable);
 typedef void CleverTapOnKVDataChangedHandler(Map<String, Object?> kvData);
 
 /// Signature of callback triggered on background isolate when a user taps on a
 /// notification from killed state.
 typedef void CleverTapOnKilledStateNotificationClickedHandler(
     Map<String, dynamic> map);
+
+/// custom code templates - start
+typedef void CleverTapCustomTemplatePresentHandler(String templateName);
+typedef void CleverTapCustomTemplateCloseHandler(String templateName);
+typedef void CleverTapCustomFunctionPresentHandler(String functionName);
+/// custom code templates - end